refactor(open-mpm): split tools/ batch 1 — mod.rs/fs_reader/mcp_tools/skill_loader
Splits four oversized files in crates/open-mpm/src/tools/ to satisfy the
500-line cap (closes part of #361):

- mod.rs: extract inline test module to registry_tests.rs (683 -> 394 lines)
- fs_reader.rs -> fs_reader/ (cwd, read_file, list_dir, grep, mod+tests)
- mcp_tools.rs -> mcp_tools/ (schema, dispatch, executor, mod+tests)
- skill_loader.rs -> skill_loader/ (resolver, loader_tool, list_tool, mod+tests)

No public API surface changes; all re-exports preserved. All open-mpm tests
pass; clippy -D warnings clean; fmt clean.
